closes #1244 - checking whether file exists before deleting it in cache
git-svn-id: http://code.elgg.org/elgg/trunk@3752 36083f99-b078-4883-b0ff-0f9b5a30f544

+++ engine/lib/cache.php
@@ -389,8 +389,11 @@ class ElggFileCache extends ElggCache {
 	 */
 	public function delete($key) {
 		$dir = $this->get_variable("cache_path");
-
-		return unlink($dir.$key);
+		
+		if (file_exists($dir.$key)) {
+			return unlink($dir.$key);
+		}
+		return TRUE;
 	}
 
 	public function clear() {
